requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
yfinance
nltk
textblob
requests
jupyter
streamlit
python-dotenv